- add Pig Statuette – 1 Year - add Goat Statuette – 2 Year - add Ice Fiend – 3 Year - add Enchanted Granite Cart – 6 Year - add Glass Kiln, Glassblowing, Functions as a heat source when ignited – 7 Year - add Enchanted Sculpting Tool, Masonry – 7 Year - add Ethereal War Boar - EtherealSoulbinder Deserialize fix.

+++ Scripts/Services/Seasonal Events/ForsakenFoes/Items/EtherealSoulbinder.cs
@@ -103,7 +103,7 @@ namespace Server.Items
             base.Deserialize(reader);
             int version = reader.ReadInt();
 
-            m_SoulPoint = reader.ReadInt();
+            m_SoulPoint = reader.ReadDouble();
         }
     }
 }
